requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
sqlalchemy-libsql==0.2.0
libsql-experimental==0.1.13
pydantic[email]==2.10.6
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.0.1
python-dotenv==1.0.0
httpx==0.27.0
pyjwt==2.8.0
aiortc==1.10.1
fastrtc[vad]==0.0.18
python-decouple==3.8
cryptography==43.0.3
anthropic==0.49.0
websockets==12.0
langchain==0.3.20
langchain-core==0.3.47
numpy==2.1.3
pydub==0.25.1
google-cloud-speech==2.31.0
google-cloud-texttospeech==2.25.0
openpyxl==3.1.5
pytest==7.4.0
requests==2.31.0
typing-extensions==4.12.2
scipy==1.15.2
soundfile==0.12.1 
elevenlabs==1.54.0
langchain-anthropic==0.3.10
langgraph==0.2.76
